+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Removed
 ### Fixed
 - fix quote replacement for non‑ASCII characters and punctuation after closing quotes
+- add missing intellisense
 ### Security